# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next Olshop API",
    "description": "REST API for Next Olshop — Single Store Online Shop untuk UMKM. Digunakan oleh Flutter mobile app.\n\n## Authentication\nSemua endpoint yang membutuhkan auth menggunakan Bearer token dari `/api/auth/login`.\n\n## Response Format\n- **Success**: `{ success: true, data: T }`\n- **Error**: `{ success: false, error: string }`\n- **Pagination**: `{ success: true, data: T[], meta: { page, limit, total, totalPages } }`\n\n## Catatan\n- Harga dalam Rupiah (IDR), disimpan sebagai string decimal (`\"299000.00\"`)\n- Tanggal dalam format ISO 8601 (UTC)\n- Berat produk dalam gram",
    "version": "1.0.0",
    "contact": {
      "name": "Next Olshop",
      "url": "https://github.com/codewithbahri"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token dari POST /api/auth/login"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "nullable": true, "example": 1 },
          "name": { "type": "string", "example": "Wireless Earbuds Pro" },
          "slug": { "type": "string", "example": "wireless-earbuds-pro" },
          "description": { "type": "string", "nullable": true, "example": "High-quality wireless earbuds" },
          "price": { "type": "string", "example": "299000.00", "description": "Decimal string (IDR)" },
          "image": { "type": "string", "nullable": true, "example": "https://images.unsplash.com/..." },
          "stock": { "type": "integer", "example": 50 },
          "weight": { "type": "integer", "example": 200, "description": "Berat dalam gram" },
          "isActive": { "type": "boolean", "example": true },
          "isFeatured": { "type": "boolean", "example": false },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "category": { "$ref": "#/components/schemas/Category" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Elektronik" },
          "slug": { "type": "string", "example": "elektronik" },
          "description": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "orderNumber": { "type": "string", "example": "ORD2604ABC1" },
          "status": { "type": "string", "enum": ["waiting_payment", "packing", "shipping", "delivered", "expired", "cancelled"], "example": "waiting_payment" },
          "subtotal": { "type": "string", "example": "598000.00" },
          "shippingCost": { "type": "string", "example": "45000.00" },
          "total": { "type": "string", "example": "643000.00" },
          "shippingName": { "type": "string", "example": "John Doe" },
          "shippingPhone": { "type": "string", "example": "08123456789" },
          "shippingAddress": { "type": "string", "example": "Jl. Raya Sudirman No. 123" },
          "notes": { "type": "string", "nullable": true },
          "willExpiredAt": { "type": "string", "format": "date-time", "nullable": true },
          "paidAt": { "type": "string", "format": "date-time", "nullable": true },
          "shippingAt": { "type": "string", "format": "date-time", "nullable": true },
          "deliveredAt": { "type": "string", "format": "date-time", "nullable": true },
          "expiredAt": { "type": "string", "format": "date-time", "nullable": true },
          "cancelledAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItem" }
          },
          "invoice": { "$ref": "#/components/schemas/Invoice" },
          "shipping": { "$ref": "#/components/schemas/Shipping" }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "productId": { "type": "integer" },
          "productName": { "type": "string", "example": "Wireless Earbuds Pro" },
          "productImage": { "type": "string", "nullable": true },
          "price": { "type": "string", "example": "299000.00" },
          "quantity": { "type": "integer", "example": 2 },
          "subtotal": { "type": "string", "example": "598000.00" }
        }
      },
      "Invoice": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "xenditId": { "type": "string", "example": "inv_123456" },
          "invoiceUrl": { "type": "string", "example": "https://checkout.xendit.co/invoices/inv_123456" },
          "paymentMethod": { "type": "string", "nullable": true },
          "paymentChannel": { "type": "string", "nullable": true },
          "amount": { "type": "string", "example": "643000.00" },
          "status": { "type": "string", "enum": ["pending", "paid", "expired", "cancelled"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Shipping": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "recipientName": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "courierName": { "type": "string", "example": "J&T Regular" },
          "courierCompany": { "type": "string", "example": "jnt" },
          "courierType": { "type": "string", "example": "REG" },
          "price": { "type": "string", "example": "45000.00" },
          "trackingId": { "type": "string", "nullable": true, "example": "TRK123456" },
          "waybillId": { "type": "string", "nullable": true },
          "estimateDays": { "type": "string", "nullable": true, "example": "3-4" },
          "status": { "type": "string", "nullable": true, "example": "pending_pickup" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message" }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 150 },
          "totalPages": { "type": "integer", "example": 8 }
        }
      }
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List Products",
        "description": "Ambil daftar produk aktif dengan filter & pagination",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "Filter by category slug", "example": "elektronik" },
          { "name": "featured", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] }, "description": "Filter featured products only" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by product name" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "List produk dengan pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List Orders",
        "description": "Ambil daftar pesanan user dengan items, invoice, dan shipping",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create Order",
        "description": "Buat pesanan dari keranjang belanja. Otomatis: deduct stock, create Xendit invoice, clear cart. Response berisi paymentUrl untuk redirect ke halaman pembayaran.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["addressId", "courierName", "courierCompany", "courierType", "courierPrice"],
                "properties": {
                  "addressId": { "type": "integer", "example": 1, "description": "ID alamat pengiriman" },
                  "courierName": { "type": "string", "example": "J&T Regular", "description": "Nama kurir (dari shipping rates)" },
                  "courierCompany": { "type": "string", "example": "jnt", "description": "Kode perusahaan kurir" },
                  "courierType": { "type": "string", "example": "REG", "description": "Tipe layanan kurir" },
                  "courierPrice": { "type": "integer", "example": 45000, "description": "Harga ongkir (dari shipping rates)" },
                  "notes": { "type": "string", "example": "Tolong ketuk pintu 2x", "description": "Catatan tambahan (opsional)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order dibuat, redirect ke paymentUrl",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "allOf": [
                        { "$ref": "#/components/schemas/Order" },
                        {
                          "type": "object",
                          "properties": {
                            "paymentUrl": { "type": "string", "example": "https://checkout.xendit.co/invoices/inv_123456" }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Keranjang kosong / stok tidak cukup / validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Alamat tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/webhooks/xendit": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Xendit Payment Webhook",
        "description": "Webhook dari Xendit untuk notifikasi pembayaran. Diverifikasi via `x-callback-token` header. **Jangan panggil manual — ini untuk Xendit server.**",
        "parameters": [
          { "name": "x-callback-token", "in": "header", "required": true, "schema": { "type": "string" }, "description": "Xendit webhook verification token" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "external_id": { "type": "string", "description": "Order number" },
                  "status": { "type": "string", "enum": ["PAID", "EXPIRED"] },
                  "payment_method": { "type": "string" },
                  "payment_channel": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook processed" },
          "401": { "description": "Invalid callback token" }
        }
      }
    },
...
  }
}